Splitting the UI component
